+++ tests/test_coding/test_structures.py
@@ -10,6 +10,10 @@ from satella.coding.structures import TimeBasedHeap, Heap, typednamedtuple, \
 
 class TestTimeBasedHeap(unittest.TestCase):
 
+    def test_dict_object(self):
+        self.assertTrue(DictObject({'a': 5, 'b': 'test'}).is_valid_schema({'a': int, 'b': str}))
+        self.assertFalse(DictObject({'a': 5, 'b': 'test'}).is_valid_schema({'a': int, 'b': int}))
+
     def test_omni(self):
         class Omni(OmniHashableMixin):
             _HASH_FIELDS_TO_USE = ['a']